login error checking tehtud.

+++ src/main/java/ttt/valiit/abja_kino_back/business/genre/GenreService.java
@@ -1,23 +1,22 @@
 package ttt.valiit.abja_kino_back.business.genre;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ttt.valiit.abja_kino_back.domain.genre.Genre;
 import ttt.valiit.abja_kino_back.domain.genre.GenreRepository;
 import ttt.valiit.abja_kino_back.domain.movie.MovieMapper;
+import ttt.valiit.abja_kino_back.domain.movie.MovieRepository;
 import ttt.valiit.abja_kino_back.infrastructure.exception.GenreExistsException;
 import ttt.valiit.abja_kino_back.infrastructure.exception.ResourceNotFoundException;
 
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
 public class GenreService {
     private final GenreRepository genreRepository;
-    private final MovieMapper movieMapper;
+    private final MovieRepository movieRepository;
 
-    public GenreService(GenreRepository genreRepository, MovieMapper movieMapper) {
-        this.genreRepository = genreRepository;
-        this.movieMapper = movieMapper;
-    }
 
     public List<Genre> getAllGenres() {
 
@@ -47,6 +46,11 @@ public class GenreService {
         Genre genre = genreRepository.findById(id).orElseThrow(
                 () -> new ResourceNotFoundException("Žanri ID ei leitud")
         );
+
+        if (movieRepository.existsByGenre(id)) {
+            throw new GenreExistsException("Žanr on seotud filmiga");
+        }
+
         genreRepository.delete(genre);
